refactor: 스트림으로 메서드 개선
